moved methods to contoller and made them private; add filename field to connector
